chore: rename batch and block output files to batch_outputs and block_outputs respectively
