+++ crates/miden-protocol/src/batch/batch_outputs.rs
