+++ crates/miden-protocol/src/block/block_outputs.rs
